[OPIK-5826][TS SDK] Inline `getTestSuites` logic into client, remove helper function, and refactor project ID resolution with `resolveProjectId`.
